WIP - LLRF Monitor Window
