+++ pyqt-apps/siriushla/li_rf_llrf/__init__.py
@@ -1,1 +1,2 @@
-from .main import MainWindow
+from .monitor import LLRFMonitor
+from .main import LLRFMain
